Added basic generator syntax and basic generators.
We're using build tags to separate generator-generation code
from the actual generated-generators. Not confusing at all.

This allows to have `Generator[T]` as the return type of generators
in both generator generation, and real generator code.
This allows having it actually be different types!

The trick is that the generation sources are only built when generating
and the real code is not built when generating.