Added range adapters for slice and map.
This is essential for implementing range-based-for.
It allows us to maintain the state of iteration.

The trick here is that we need to define a new variable,
and we currently have no mechanism for it.
But since we know all the properties of this variable,
we can add a "gengen vars" field and just place it there
as a string instead of generating it in a smarter way.