Fix some weird issue with calling qualified names from packages.
The package names would not show up in the variable lookup, and so
using them in selector expressions would leave an empty heading.

This will also happen with global variables & constants - we don't
walk the top-level, so we don't define any globals.
We probably need to do a walk for that.

So the idea is to do 2 walks.
One for rendering everything, and one just for collecting global
definitions.