Better code example for the talk- use the library example instead of the paginated API.
This shows how tricky iterators can be.
We work from the inside out instead of outside in
And we loops despite only trying to get a single value.

You may have a better solution for that - but that's exactly the point.
Implementing iterators should not be an intellectual challenge.
It should be easy and straightforward.